// NutriTEC/nutritec.client/src/app/interfaces/RegisterNutritionist.ts
export interface RegisterNutritionist {
  nombre: string;
  apellido1: string;
  apellido2: string;
  nacimiento: Date;
  cedula: string;
  codnutricionista: string;
  peso: number;
  IMC: number;
  direccion: string;
  numerotarjeta: string;
  tipocobro: string;
  email: string;
  password: string;
}